merge

+++ db/migrate/20170407201558_create_abilities.rb
@@ -3,7 +3,7 @@ class CreateAbilities < ActiveRecord::Migration[5.0]
     create_table :abilities do |t|
     	t.string :ability_type
       t.integer :value
-      t.float :score
+      t.integer :score
       t.references :character_stat
     end
   end
